% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIFplus-package.R
\docType{package}
\name{RIFplus-package}
\alias{RIFplus-package}
\title{RIFplus}
\description{
Blah blah blah
}
\details{
\tabular{ll}{ Package: \tab RIFplus\cr Type: \tab Package\cr Version:
\tab 0.1.0\cr Date: \tab 2016-12-19\cr License: \tab GPL (>=3)\cr LazyLoad:
\tab yes\cr }
}
\author{
Andrea Rau, Toni Reverter
}

