% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIFplus.R
\name{RIFplus}
\alias{RIFplus}
\title{RIFplus function}
\usage{
RIFplus(exprs, TFnames, conds)
}
\arguments{
\item{exprs}{Blah blah}

\item{TFnames}{Blah blah}

\item{conds}{Blah blah}
}
\value{
Value
}
\description{
RIFplus function
}

